manager後台CRUD
