+++ app/controllers/backstage/managers_controller.rb
@@ -1,0 +1,7 @@
+class Backstage::ManagersController < ApplicationController
+  before_action :authenticate_manager!
+
+  def show
+    @restaurants = current_manager.restaurants
+  end
+end
