--- app/controllers/managers_controller.rb
@@ -1,5 +1,0 @@
-class ManagersController < ApplicationController
-  def show
-    @restaurants = current_manager.restaurants
-  end
-end
