+++ app/helpers/backstage/managers_helper.rb
@@ -1,0 +1,2 @@
+module Backstage::ManagersHelper
+end
