+++ app/helpers/backstage/restaurants_helper.rb
@@ -1,0 +1,2 @@
+module Backstage::RestaurantsHelper
+end
