+++ app/javascript/stylesheets/backstagea_pplication.scss
@@ -1,0 +1,49 @@
+@import "tailwindcss/base";
+@import "tailwindcss/components";
+@import "tailwindcss/utilities";
+
+@layer base{
+  h1,h2,h3,h4,h5,h6,p,span{
+    @apply text-gray-800;
+  }
+}
+
+@layer components{
+  .input-border-click{
+    @apply  focus:ring-major
+            focus:border-major
+              focus:outline-none
+              border
+              rounded 
+  }
+}
+
+@layer components{
+  .input-box{
+    @apply  focus:ring-major
+            focus:border-major
+              focus:outline-none
+              border
+              rounded
+              mb-2 
+  }
+}
+
+@layer components{
+  .home-img{
+    @apply pb-8 border-b-2
+  }
+
+  .major-btn{
+    @apply border text-sm bg-major text-white rounded py-1 px-2 mr-2 hover:bg-yellow-600
+  }
+
+  .member-system-flex{
+    @apply max-w-screen-lg py-10 m-auto mt-24 border rounded md:flex md:justify-between px-14
+  }
+}
+
+.date{
+  padding: 2px;
+} 
+
