+++ app/views/backstage/managers/show.html.erb
@@ -1,0 +1,12 @@
+<h1><%= current_manager.name %></h1>
+<%= link_to "+ 新增餐廳", new_backstage_manager_restaurant_path(current_manager.id), class: "major-btn" %>
+
+<% @restaurants.each do |restaurant| %>
+  <div>
+    <%= link_to restaurant.title, backstage_restaurant_path(restaurant) %>
+    <%= restaurant.tel %>
+    <%= restaurant.address %>
+    <%= link_to "編輯餐廳資訊", edit_backstage_restaurant_path(restaurant), class: "major-btn"%>
+    <%= link_to "刪除", backstage_restaurant_path(restaurant), method: "delete", class: "major-btn"%>
+  </div>
+<% end %>
