+++ app/views/backstage/restaurants/_error.html.erb
@@ -1,0 +1,7 @@
+<% if (@restaurant.errors.any?) %>
+  <ul>
+    <%= @restaurant.errors.full_messages.each do |message| %>
+      <li><%= message %></li>
+    <% end %>
+  </ul>
+<% end %>
