+++ app/views/backstage/restaurants/new.html.erb
@@ -1,0 +1,5 @@
+<h1>新增餐廳</h1>
+
+<%= render "error", restaurant: @restaurant %>
+
+<%= render 'form', restaurant: @restaurant, url: backstage_manager_restaurants_path(current_manager.id) %>
