+++ app/views/backstage/restaurants/show.html.erb
@@ -1,0 +1,2 @@
+<h1><%= @restaurant.title %></h1>
+<%= link_to "編輯座位資料" %>
