--- app/views/restaurants/new.html.erb
@@ -1,28 +1,0 @@
-<h1>新增餐廳</h1>
-
-<% if (@restaurant.errors.any?) %>
-  <ul>
-    <%= @restaurant.errors.full_messages.each do |message| %>
-      <li><%= message %></li>
-    <% end %>
-  </ul>
-<% end %>
-
-<%= form_with(model: @restaurant, url: manager_restaurants_path) do |f| %>
-  <div>
-    <%= f.label :title, "餐廳名稱：" %>
-    <%= f.text_field :title, class: "input-box" %>
-  </div>
-
-  <div>
-    <%= f.label :address, "餐廳地址：" %>
-    <%= f.text_field :address, class: "input-box" %>
-  </div>
-
-  <div>
-    <%= f.label :tel, "餐廳電話：" %>
-    <%= f.text_field :tel, class: "input-box" %>
-  </div>
-  <%= f.submit "新增餐廳" ,class: "major-btn" %>
-<% end %>
-
