--- app/views/restaurants/show.html.erb
@@ -1,2 +1,0 @@
-<h1><%= @restaurant.title %></h1>
-<%= link_to "編輯座位資料", edit_restaurant_path, class: "major-btn" %>
