+++ config/routes.rb
@@ -1,9 +1,12 @@
 Rails.application.routes.draw do
   devise_for :managers
 
-  resources :managers, only: [:show] do
-    resources :restaurants, shallow: true, only: [:new, :create, :show, :edit, :update, :destroy]
+  namespace :backstage do
+    resources :managers, only: [:show] do
+      resources :restaurants, shallow: true, only: [:new, :create, :show, :edit, :update, :destroy]   
+    end
+    root 'managers#show'
   end
-  
+
   root 'home#index'
 end
