+++ test/controllers/backstage/restaurants_controller_test.rb
@@ -1,0 +1,7 @@
+require "test_helper"
+
+class Backstage::RestaurantsControllerTest < ActionDispatch::IntegrationTest
+  # test "the truth" do
+  #   assert true
+  # end
+end
